requirements.txt
diskcache>=5.6.3
google-generativeai>=0.3.2
python-dotenv>=1.0.0 